feat: refactor backstage code
