+++ package.json
@@ -34,10 +34,9 @@
     "graphql": "^14.1.1",
     "graphql-tools": "^4.0.4",
     "json-server": "^0.14.2",
-    "lodash": "^4.17.11",
+    "lodash": "^4.17.21",
     "log4js": "^4.0.2",
-    "moment": "^2.27.0",
-    "node-cache": "^5.1.2",
+    "moment": "^2.29.1",
     "pg": "^7.7.1",
     "querystring": "^0.2.0",
     "rimraf": "^2.6.2"
