+++ src/docs/GenerateFileSchema.js
@@ -1,4 +1,4 @@
 const { writeFileSync } = require('fs');
-const { typeDefs } = require('../graphql/Schema');
+const { typeDefs } = require('../Schema');
 
 writeFileSync(`${__dirname}/schemaDoc.graphql`, (typeDefs));
