--- src/graphql/common/TypeDefs.js
@@ -1,10 +1,0 @@
-const TypeDefs = [`
-    type MapStringToString {
-     #key
-     key: String
-     #value
-     value: String
-  }
-`];
-
-module.exports = TypeDefs;
