+++ src/operations/authentication/helpers/AxiosUtils.js
@@ -1,4 +1,4 @@
-const config = require('../../config');
+const config = require('../../../config');
 
 class AxiosUtils {
   static get GET() {
