+++ src/operations/configuration/Resolvers.js
@@ -1,0 +1,13 @@
+const getConfig = require('./query.getConfig')
+const updateConfig = require('./mutation.updateConfig')
+
+const Resolvers = {
+  Query: {
+    getConfig
+  },
+  Mutation: {
+    updateConfig
+  }
+};
+
+module.exports = Resolvers;
