+++ src/operations/device/Resolvers.js
@@ -1,0 +1,13 @@
+const getDeviceById = require('./query.getDeviceById')
+const getDevices = require('./query.getDevices')
+const getDeviceHistoryForDashboard = require('./query.getDeviceHistoryForDashboard')
+
+const Resolvers = {
+  Query: {
+    getDeviceById,
+    getDevices,
+    getDeviceHistoryForDashboard
+  },
+};
+
+module.exports = Resolvers;
