+++ src/operations/device/TypeDefs.js
